Extract test code to simplify TestApp file
